Fix hex32dec
hex32dec caused problems of type, since it was a uint.
Changed to double

+++ matlab/teensy/hex32dec.m
@@ -1,5 +1,5 @@
 function [ num ] = hex32dec( data )
 %HEX32DEC Transform hex repr of 32 bit int to 32 bit int
-    num = typecast(uint32(hex2dec(data)), 'int32');
+    num = double(typecast(uint32(hex2dec(data)), 'int32'));
 end
